+++ CHANGES.rst
@@ -46,6 +46,7 @@ Changelog
 
 **Removed**
 
+- #1149 Remove "Unpublished" filter from Analysis Requests listing
 - #1132 Remove "Submitted by current user" icon from AR listing (performance)
 - #1125 Remove Sample views, listings and links to Sample(s) from everywhere
 - #1118 Removed all legacy Bika Listing / Advanced Filtering from Codebase